refactor: refactor job for sending course notifications
